move GraphTile::GetTileId to GraphId::FromTilPath as a static factory ctor
